INTERLOK-851 #time 2h
Turns out from logging that we must have a groupID specified; so now we
default to a unique-id if no groupid is specified.
However, the embedded Kafka tests are not working based on the logging as we
get a NO_LEADER exception, which I can't yet work out where that's coming from.